fix: target labels cast long()

+++ lib/smooth_cross_entropy.py
@@ -11,7 +11,7 @@ class SmoothCrossEntropy(nn.Module):
     def forward(self, logits, labels):
         num_classes = logits.shape[-1]
         alpha_div_k = self.alpha / num_classes
-        target_probs = F.one_hot(labels, num_classes=num_classes).float() * \
+        target_probs = F.one_hot(labels.long(), num_classes=num_classes).float() * \
             (1. - self.alpha) + alpha_div_k
         loss = -(target_probs * torch.log_softmax(logits, dim=-1)).sum(dim=-1)
         return loss.mean()
